Replace torch and torchvision
