+++ datumaro/plugins/openvino_plugin/samples/clip_ViT-B_32_interp.py
